+++ Arena/ArenaHelpers.cs
@@ -264,13 +264,34 @@ namespace RainMeadow
                 action.Invoke(Regex.Split(array[i], "<msuB>"));
         }
 
+        public static bool CheckSameTeam(OnlinePlayer? A, OnlinePlayer? B)
+        {
+            if (A is not null && B is not null)
+            {
+                if (
+                    OnlineManager
+                        .lobby.clientSettings.TryGetValue(A, out var cA) && cA.TryGetData<ArenaTeamClientSettings>(out var tb1)
+                )
+                {
+                    if (
+                        OnlineManager
+                            .lobby.clientSettings.TryGetValue(B, out var cB) && cB.TryGetData<ArenaTeamClientSettings>(out var tb2)
+                    )
+                    {
+                        return tb1.team == tb2.team;
+                    }
+                }
+            }
+            return false;
+        }
+
         public static bool CheckSameTeam(
-            ArenaOnlineGameMode arena,
-            OnlinePlayer? A,
-            OnlinePlayer? B,
-            Creature creature,
-            Creature friend
-        )
+    ArenaOnlineGameMode arena,
+    OnlinePlayer? A,
+    OnlinePlayer? B,
+    Creature creature,
+    Creature friend
+)
         {
             if (A is not null && B is not null)
             {