+++ Game/RainMeadow.PlayerHooks.cs
@@ -2260,12 +2260,36 @@ public partial class RainMeadow
             c.Emit(OpCodes.Brtrue, skip);
             c.Index += 6;
             c.MarkLabel(skip);
+
+
+            // reset and scan
+            c.Index = 0;
+            // looking for the call to get_Value() from the cfgArtificerExplosionCapacity config
+            while (c.TryGotoNext(MoveType.After,
+                i => i.MatchLdsfld<MoreSlugcats.MoreSlugcats>(nameof(MoreSlugcats.MoreSlugcats.cfgArtificerExplosionCapacity))
+            ))
+            {
+                // ignore type definition just override the value
+                if (c.TryGotoNext(MoveType.After, i => i.MatchCallOrCallvirt(out _)))
+                {
+                    c.Emit(OpCodes.Ldarg_0);
+                    c.EmitDelegate<Func<int, Player, int>>((originalValue, self) =>
+                    {
+                        if (RainMeadow.isArenaMode(out var arena))
+                        {
+                            return arena.artiExplosionCount;
+                        }
+                        return originalValue;
+                    });
+                }
+            }
         }
         catch (Exception e)
         {
             Logger.LogError(e);
         }
     }
+
 
     private bool Player_CanMaulCreature(On.Player.orig_CanMaulCreature orig, Player self, Creature crit)
     {