+++ GameModes/ArenaCompetitiveGameMode.cs
@@ -62,6 +62,8 @@ namespace RainMeadow
         public int denScore = RainMeadow.rainMeadowOptions.ArenaDenScore.Value;
 
         public int emptyKillTagScore = RainMeadow.rainMeadowOptions.ArenaDenScore.Value;
+
+        public bool winByScore => spearScore > 0 || aliveScore > 0 || emptyKillTagScore > 0;
         public bool challengeDenEjection = RainMeadow.rainMeadowOptions.ChallengeDenEjection.Value;
 
         public string paincatName;
@@ -126,6 +128,8 @@ namespace RainMeadow
             .rainMeadowOptions
             .ArenaSaintAscendanceTimer
             .Value;
+
+        public int artiExplosionCount = MoreSlugcats.MoreSlugcats.cfgArtificerExplosionCapacity.Value;
         public int watcherCamoTimer = RainMeadow.rainMeadowOptions.ArenaWatcherCamoTimer.Value;
         public int watcherRippleLevel = RainMeadow.rainMeadowOptions.ArenaWatcherRippleLevel.Value;
         public int amoebaDuration = RainMeadow.rainMeadowOptions.AmoebaDuration.Value;